disk: move Filesystem and methods to separate file
Split up disk.go into smaller more managable pieces: continue
with moving `Filesystem` to its own file.
